kafka-sinks: insist on reading progress messages known to exist
When we fetch the messages of the progress topic we do so with a 10
second timeout which results the code interprets as "no more messages
available". If we've only read a prefix of the topic and hit a timeout
it is possible to successfully return a stale `latest_ts`.

This commit adds an additional check that verifies we've read at least
until the high watermark we obtained during startup.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>