balancerd: fetch tenant id directly from password
Previously we would contact frontegg to get a tenant id so we could
correctly route incoming connections. However the tenant is already
present in the password. Use that to avoid needing to make any
connection to frontegg.

Before this change, the balancer would act as a pre-auth gateway that
would prevent invalid auth'd sessions from ever getting to envd. That's
probably good. But it came at the expense of double connecting to
frontegg (since envd does its own auth), which is bad because it would
allow us to hit our quota limits in frontegg easier. But the very bad
part is that just by knowing the region address for a balancerd (which
perhaps won't be that secret), an attacker could very easily send lots
of traffic with bogus tenant ids and we would happily send all of those
to frontegg. This could have resulted in us being DOS'able due to
frontegg quotas.

The change here makes it so that the DNS lookup is done first, and then
auth still happens in envd. This means that in order to convince us to
make an API request to frontegg, you first must show up with a known
tenant id, a much higher bar.

In the future we could get to a point where balancer checkes the tenant
id, does the auth, then forwards the JWT to envd. This would offload all
initial connection and auth handling into the balancer layer and out of
envd.