catalog: Listen for changes in persist catalog (#23274)
This commit updates the persist catalog so that it is able to listen
for updates made by other catalogs. Currently, the functionality is
only used by read-only catalogs because they need to co-exist with
writable catalogs and be able to see their writes. This commit lays the
foundation for future work needed to expose catalog updates upstream to
the adapter layer.

Works towards resolving #22392
Works towards resolving #22481