mzcompose: Fix running multiple commands at once
in multiple threads. On Linux this happened to work, but on macOS for
example `bin/mzcompose --find sqlsmith down && bin/mzcompose --find
sqlsmith run default` would fail with cut off empty /dev/fd/3 files.

My understanding of the issue is that the old behaviour was always prone
to such issues, it was just never noticed because only one thread was
running compositions at a time.