storage: validate replication slot before downgrading capabilities
The code on `main` had a bug in the `replication` operator of the
postgres source where it was first determining the current slot LSN,
then downgrading its capabilities to that LSN, and then checking if the
LSN has overcompacted. This order is incorrect because in the event that
the slot has overcompacted we've already made final statements about the
contents of the collection until the slot LSN which may be incorrect.

This commit fixes this issue by only touching the capabilities as a
result of receiving a `PrimaryKeepAlive` message from the upstream,
which can only happen if the overall resume frontier requirements can be
serviced by the slot.

Additionally, this commit changes the `assert` into a normal transient
error so that occurences of it lead to just that dataflow restarting
instead of the whole cluster.

Fixes #23287

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>