storage: guard against build_fallible footgun
Because the capabilities that we retained were stored in a btree map
which was part of the future state, when we encountered an error and
returned from the future those were dropped and created bogus frontiers
that propagated down the dataflow. This can cause incorrect data to be
persisted. Until `build_fallible` is fixed we must implement this ugly
workaround.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>