Record the heap capacity of the shared row in compute
This change adds support to logging the heap capacity of the shared row in
compute (but not other uses of the shared row.) The heap capacity should
be low, but depends on the largest row we ever processed. This metric gives
us the necessary insights to introducing a policy about sizing-down the
row.

Fixes #23150.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>