storage: better separate encoding from publishing in Kafka sinks
Previously the two operators involved in Kafka sinks (`encode_stream`
and `produce_to_kafka`) had overlapping responsibilities. The former
created the progress and data topics and published schemas to the schema
registry, while the latter read the progress topic and wrote to the data
topic. The two operators were scheduled concurrently, meaning that the
former operator would be creating the topics as the latter operator was
attempting to read from/write to them.

This commit moves all responsibility for managing the Kafka topics to
the `produce_to_kafka` operator, leaving `encode_stream` with only the
responsibility of publishing the schemas to the schema registry. The new
separation of responsibilities makes the code much easier to follow.

Two nice side effects:

  * The ugly `KafkaSinkAvroFormatState` enum disappears.
  * The progress topic is only read a single time.

There is much more potential for refactoring, but I didn't want to fight
too hard with @sploiselle's work in #23357.