explain: fix `OptimizerNotice` rendering
At the moment user-facing notices are rendered using the column
reference humanizer scheme introduced by #22959 wich is unambigous, but
not suitable for copy-pasting. For notices we want to stick to the mode
from #21463.

This commit extends the `HumanizedExpr` trait with a generic parameter
that is bound by a new `HumanizerMode` trait. We add two types that
implement that trait:

- `HumanizedExplain` for expressions that need to be humanized when
  rendering an `EXPLAIN` result.
- `HumanizedNotice` for expressions that need to be humanized
   when rendering an `OptimizerNotice`.