storage: fix kafka transational fencing logic
Until now we were putting the `worker_id` in our Kafka
`transactional.id`. For each dataflow instantiation we pick a single
worker that will be responsible for talking with Kafka. The worker
picked is the one whose `worker_id` satisfies `(sink_id.hashed() %
total_workers) == worker_id`.

The problem is that the chosen worker can change if the cluster gets
resized or if the hashing implementation changes between two releases.
This means a zombie cluster will not necessarily be fenced out by the
new one, leading us to produce duplicate values, violating our exactly
once guarantees.

This commit fixes the problem by transitioning our `transactional.id` to
always pretend to be worker 0 and includes a defensive fencing of the
old style `transactional.id` to minimize the probability of zombies not
being fenced.

We will eventually transition to an even better `transactional.id`
scheme[1] but that requires a more involved migration.

[1] https://github.com/MaterializeInc/materialize/issues/23198

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>