catalog: Make the persists catalog schema binary
This commit updates the schema of the persist shard that the persist
catalog uses such that the data is represented as raw binary instead of
a strongly typed Rust struct. This allows us to read/write the catalog
data as raw binary instead of the StateUpdateKind enum. When executing
an upgrade, the current StateUpdateKind may not be able to deserialize
the contents of the persist shard which may be from an earlier version.
Additionally, when executing an upgrade, we need to read the data
using one version, retract the data using the same version, and insert
the data using another version. All this is straightforward to do if
we're reading/writing raw binary, but very tricky to have a single Rust
struct representing all different versions. Uses of the persist shard
outside or upgrades need to manually deserialize the data after reading
it in.

Works towards resolving: #22977