catalog: Make persist catalog generic over key
This commits updates the persist catalog's `PersistHandle` so that it's
generic over the persist shard key type. This will allow us to
read/write from/to persist with either a strongly typed Rust struct or
raw binary. This is required for catalog migrations which may not have
a single rust struct that can be used for both retractions and
insertions.

Works towards resolving: #22977