transform: make `DataflowMetainfo` notice type generic
The type of the notices stored in the `optimizer_notices` field depends
on the execution stage - within optimizer passes we still need to use
the `RawOptimizerNotice` type, but before storing the final result in
the `CatalogPlans` struct we need to move to a `OptimizerNotice` vector
with a `CatalogState::render_notices` call.

We set `RawOptimizerNotice` as the default `Notice` type in order to
minimize the diff.