pgwire: stop eagerly sending `ReadyForQuery`
In #16949, we adjusted connection startup to send the initial
`ReadyForQuery` message *before* the coordinator had actually handled
the registration of the session. We did this to better meet our SLA
guarantees around connection startup.

There are several downsides to this behavior, however:

  1. Startup notices aren't sent until the first query. This is
     particularly bad for the proposed welcome notice proposed in #23440.
  2. Unknown username errors aren't generated until after the first query
     is submitted.
  3. Query latency for the first query is often much worse than it would
     be otherwise, as processing of the first query is blocked on
     session registration, which requires a system table write. This is
     particularly bad for fast index lookup queries that wouldn't
     otherwise need to hit `persist`.

This commit removes the behavior, restoring the old behavior of waiting
for session registration before sending the initial `ReadyForQuery`
message

The primary motiviation is eliminating downside (3), which we weren't
aware of when we implemented the eager `ReadyForQuery` behavior. Our
users expect consistently low response times to queries, and sacrificing
query response time in order to improve our ability to hit our uptime
SLA does not seem to be the right tradeoff--especially since we are
comfortably meeting our uptime SLA, even when using a strict definition
of "uptime" that means "SHOW VIEWS succeeds" rather than "connection
requests succeed.".