Use ColKeySpine types in RowUnit arrangements
This commit improves the representation of arrangements that are
specialized to empty values by employing types based on ColKeySpine
instead of ColValSpine. The former types eliminate the layer
representing values, thus achieving lower sizes in expectation. At the
moment, there is a transition happening where ColValSpine types are
already in ord_neu while ColKeySpine types are not. This may create
unusual situations where improving representation leads to worse
outcomes due to lack of deduplication in arrangements. We expect the
latter to be resolved, however, when ColKeySpine also becomes based on
ord_neu.