sql-parser: add parens around options in CREATE CLUSTER
For historical reasons, the syntax for CREATE CLUSTER and CREATE CLUSTER
REPLICA does not require nor permit surrounding the options with
parenthesis. This is at variance with how options are represented in all
other SQL statements (e.g., `CREATE CONNECTION`, `ALTER ... SET`).

It's too late to *mandate* the use of parentheses around the options,
but we can at least *allow* the use of parentheses, and update the docs
to use the consistent form of the syntax.