storage-client: fix liveness issue in determine_latest_progress_record
Before this commit, Kafka sinks would occasionally get wedged on boot
inside of the `determine_latest_progress_record` function. The issue was
particularly prevalent when there were multiple other live sinks writing
to the progress topic.

The issue was introduced in by a refactoring in PR #23300. Funnily
enough, @bkirwi even called out a liveness issue during code review, but
this one slipped through. This liveness issue is very similar, if not
identical, to one that @bkirwi fixed several months ago in #19879.

The cause was that the message fetching loop in
`determine_latest_progress_record` would `continue` upon discovering a
progress message with a key for a different sink, without checking if
the consumer had reached the the high water mark and breaking out of the
loop if so. The only way to hit the high water mark check was if the
*last* progress record in the topic was for the sink under
consideration. In the case of a live, high traffic progress topic, this
is unlikely; in the case of a new sink, it's altogether impossible.

The upshot is we were always relying on Kafka returning `PartitionEOF`
(or timing out) to break out of the loop. But as @bkirwi described in
his writeup for #19879, `PartitionEOF` messages are not guaranteed to be
delivered, and tend to get dropped on high-traffic topics.

This commit adjusts `determine_latest_progress_record` to restore the
old and correct behavior of checking whether the high water mark has
been reached after every message read from the progress topic.

It also substantially refactors the method to improve clarity:

  * Many documentation comments are added; in particular @bkirwi's
    argument for the safety of the overall approach from #19879.
  * Comparing the consumer's position to the high water mark is moved to
    the loop condition, and reading a message from the topic is moved to
    the body. This better captures the intent of the loop--we need to
    keep reading messages until we reach the high water mark--and means
    we don't need to check the consumer's position after we exit the
    loop.
  * `get_next_message` is folded into `get_latest_ts`, so that failure
    to read a message can be propagated as the error that it is.
  * The loop tracks positions, rather than offsets. This avoids
    `Option`s (the initial position is 0).
  * Polling for progress records uses a more sensible timeout that's
    based on the transaction idle timeout rather than the default
    metadata fetch timeout.
  * The retry loop around `get_latest_ts` is removed, as the comment on
    the retry policy was wrong (all errors were retried, not just
    transient errors), and we can instead just fail and rely on the
    outermost `SuspendAndResume` loop to restart the sink.