storage-client: enable compaction of progress topics
When Materialize creates a progress topic, enable key-based compaction
on the topic so that progress records do not pile up forever. Also add
some documentation that points users in the right direction if they
choose to manage the creation of topics themselves.

As a side effect, this drops support for Redpanda versions before v22.2,
as those versions did not support transactional operation on compacted
topics.