adapter,storage: plumb env ID via ConnectionContext, not sources/sinks
This commit refactors the adapter and storage layers to plumb the
environment ID from environmentd to clusterd via the ConnectionContext
type, rather than via individual source/sink configuration structs.
Semantically, the environment ID is fixed global state for environmentd
and clusterd, and so it belongs in a struct representing global state.
There should be no behavior changes as a result of this refactoring.

The motivation is standardizing how names for Kafka resources are
created. Previously, the storage-types package was responsible for
creating group ID names for Kafka sources (Kafka sources threaded an
environment ID through their configuration), but the adapter package was
responsible for creating progress topic names for Kafka sinks (Kafka
sinks did not thread an environment ID through their configuration).

This commit appoints storage-types the blessed location for generating
names for Kafka resources, and having access to the environment ID in
the ConnectionContext is what makes it possible.