catalog: Fix persist update apply (#23369)
This commit fixes the persist catalogs logic for applying updates.
Previously, the catalog would apply updates within a single timestamp in
whatever order they arrived, which in general is not correct. For
example imagine the following two updates coming in at a single
timestamp:

  a. key, old_value, ts, -1
  b. key, new_value, ts, +1

Applying them in the order (a) then (b) is correct, we first remove the
old value from our in-memory map and then insert the new value. However,
applying them in the order (b) then (a) is not correct, we first
overwrite the old value with the new value in our in-memory map and then
remove the new value. We must apply the retractions before the
insertions.

Furthermore, if updates in a single timestamp are not consolidated, then
it's not enough to apply the retractions before the insertions. For
example, imagine the following updates coming in at a single timestamp:

  a. key, value1, ts, -1
  b. key, value2, ts, +1
  c. key, value2, ts, -1
  d. key, value3, ts, +1

Applying them in the order (a), (b), (c), (d) is correct, we end up with
the value value3 in our in-memory map. The order (a), (c), (d), (b) is
not correct, even though it applies all retractions before insertions,
we end up with value value2 in our in-memory map.

The persist apply logic happened to work by accident because catalog
transactions only ever involve a single retraction and a single
insertion. Also, persist happen to give us retractions before
insertions.

This commit fixes the apply logic by consolidating the updates,
guaranteeing that per (key, ts) pair there's a maximum of a single
retraction and a single insertion. Additionally, it sorts the updates by
diff within a timestamp to ensure that retractions are before
insertions.

Works towards resolving #22977