storage: detect invalid replication slots
Adds detection for the case where the replication slot becomes invalid
due to the `max_slot_keep_wal` parameter. When this happens there is no
action we can take to recover the source, therefore we produce a
definite error and immediately advance to the empty antichain, forever
erroring the source.