persist-txn: fix bug resulting in phantom latest_write
This fixes the regression test I previously added for a bug where:

- We commit a txn at some `t0`
- We later apply it and tidy/retract it at `t3`
- We start reading the txns shard as_of `t1` to serve a snapshot at `t2`
- `t0` < `t1` < `t2` < `t3`

The txns shard snapshot sees the write at `t0` advanced to `t1` and so
the DataSnapshot gets a `latest_write` of `Some(t1)`, but this is
incorrect (it should be `None` because we wouldn't compact to `t1`
unless we had applied everything before it). The apply will advance the
physical upper to `t0` but the `unblock_reads` call will wait forever
for it to get to `t1`.

The fix: encode the original `commit_ts` in the persist `K, V` where it
won't get advanced by compaction. I don't love this, it seems redundant
and clunky, but I've let this sit for a couple weeks and haven't come up
with anything else.

This commit also does the same for the `register_ts`, which is not
necessary to fix the bug, but only doing one would be subtle. Also,
reasoning about the compaction timestamps has been reasonably confusing
while working on persist-txn, so this is all probably better in the long
run, anyway.