catalog: Execute fence in separate step (#23355)
Previously, when a persist catalog was opened, it would right the new
epoch, and effectively fence all previous catalogs, as part of the first
catalog transaction. For uninitialized catalogs, this was part of the
initialize transaction. For initialized catalogs, this was part of
setting the deployment generation.

This commit, updates how the persist catalog is opened so that writing
the new epoch is its own separate atomic step that is done before the
first catalog transaction. This allows us to simplify catalog
transactions at the cost of complicating catalog opening. These
semantics also more closely match the stash catalog and are needed as a
prerequisites for persist catalog data migrations.

Works towards resolving #22977