persist: demo token dropping bug
@petrosagg and I examined a number of failures that occur in
proximity to an `ALTER CONNECTION` statement executing that restarts
a sink. @guswynn triaged a similar bug and, it turned out, came to
the same conclusion independently.

The common thread of these failures is that the sink's write frontier
advances to [], even though the item that it's sinking has not been
dropped.

In reading the logs and the code, it appears as if the tokens created
in `shard_source_descs` are not managed properly. It is possible for
n-1 workers to drop their tokens, and the nth worker to still be
running and see the effects of the other workers' token drops, i.e.
it receives an `Event::Progress` whose data is the empty frontier.

This commit substantiates the above claim by ensuring:
- Different workers are responsible for producing values to Kafka and
  reading values from persist.
- The Kafka-writing worker ignores any commands to restart the sink.
  This ensures that any effects of the n-1 workers dropping their
  tokens are observed by a worker.

We consider the claim substantiated if the running worker receives
an `Event::Progress` whose data is the empty frontier. This is easy
to check in the logs, because we will output:

> advancing write frontier to empty

Deterministically, using the code in this commit, running the
schema-registry-ssl-basic test in test/kafka-auth demonstrates that
the write frontier is advanced to the empty antichain, as demonstrated
by the presence of the above line in the logs.

e.g.

cd test/kafka-auth && ./mzcompose run default schema-registry-ssl-basic