ci: add shortlink for docs
We now link to the root docs page from the database code, so we need an
evergreen shortlink. Unlikely we'll ever move the docs from /docs, as
we'd break many links on the internet, but I want to have an absolute
policy that every link that is hardcoded into the binary is a shortlink,
so that we can update where they point without a new release of
Materialize if necessary.