LSP: completion capability (#23430)
This PR implements an initial completion feature for the LSP server. To
use this feature, the client needs to send the schema to the server, and
the server will build the completion items. The server only builds
completion items for two keywords (`SELECT` and `FROM`.) The completion
items returned for these keywords are:

* `SELECT`: Returns all the available columns in the schema.
* `FROM`: Returns all the available objects (tables, views, materialized
views, sources and sinks) available in the schema.

If the keyword is not one of those mentioned, the server returns nothing
to avoid confusion. Functions and snippets were not included in this PR
to make it simpler and smaller.