Fortify introspection-sources.td
This commit fortifies introspection-sources.td due to flakiness in
outcomes introduced by eventually consistent arrangement size
measurements. The size checks in the test are weakened to allow for a
binary order-of-magnitude range in acceptance of size and capacity
measurements. The range is centered around the previous
back-of-the-envelope size estimates in the test.