storage: ban custom shutdown tokens
Dataflow shutdown has proven to be extremely tricky and the only safe to
use API currently is the `AsyncOperatorBuilder` together with its button
implementation. In light of the recent shutdown bugs this PR bans all
custom tokens from storage by requiring that all tokens are instances of
the `PressOnDropButton` type which is provided by the `mz_timely_util`
crate.

Truly custom shutdown workflows can still be built by manually
manipulating the button but at the time of writing all operators are
accomodated by the default behavior.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>