persist: coordinated shutdown of shard_source_descs
The `shard_source_descs` operator wants to gracefully shutdown by waiting
for all outstanding parts to finish fetching and return their lease to
the `Subscribe` handle. This didn't quite fit with the out of the box
shutdown behavior of `AsyncOperatorBuilder` so the operator used a
custom scheme together with complicated `select!` dances.

This PR reorganizes the `shard_source_desc` operator into two separate
operators:
1. `shard_source_descs_return`: This operator is responsible for
   returning the leases of the parts and does **not** use a shutdown
   button at all (i.e it exits naturally when the frontier reaches the
   empty antichain).
2. `shard_source_descs`: This operator produces parts to be fetched and
   is a normal async operator that will get dropped when all workers
   have pressed their buttons.

The piece that makes this split work is that a shared reference to the
`Subscribe` handle is sent from the second to the first operator before
submitting any parts for fetching. This disconnects the lifetime of the
handle from the lifetime of the second operator and allows the first one
to keep it alive until its done returning parts while the second one is
dropped.

An added benefit of this split is that the complicated token/async
`select!` logic is gone.

Fixes: #23507

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>