adapter: fix timestamp oracle tests
The tests for the postgres/crdb oracle and the batching oracle both use
an external crdb (passed in via `COCKROACH_URL`). And both try and do
best-effort cleanup where they remove the `timestamp_oracle` table. This
means it can happen that one test removes the table from under another
test that is still running. The tests don't clash otherwise because we
pick UUIDs for the timeline names.

This commit fixes the issue by removing the best-effort cleanup.

Another solution would be to use individual table names but it's harder
to pass a runtime table name to the oracle implementations. Though it's
not at all impossible and we could do that if we want to invest more
time. This proposal here is the quickest fix, and if we throw away the
crdb instance after tests in CI it seems fine to do it like that for
now.

Fixes #23521