catalog: Remove usages of term Stash
This commit removes usages of the term Stash from comments, logs,
strings, and variable names and replaces it with something more generic
like "catalog" or "catalog storage". We are working on replacing the
Stash as the backing durable store for the catalog with persist, which
will make these usages inaccurate.

Works towards resolving #22392