persist-txn: support "eager" physical upper advancement on data shards
When a txn commits, the "logical upper" of each data shard is advanced
but the "physical upper" is only advanced for the data shards involved
in the txn. This creates a "lazy upper" situation that is addressed on
the read side.

This commit introduces the ability to instead advance the upper of all
registered data shards, which means we can use all totally normal
unchanged read paths. The eager uppers of course don't get the
efficiency and cost benefits of lazy uppers, but they'll allow us to
derisk the rollout to prod by breaking it into two pieces.