Remove dodgy activity log truncation logic
This blew up a lot sooner than I thought it would,
due to an error in my back-of-the-envelope calculations.

I now believe that we should leave truncation off and just let these
relations grow forever. However, we don't need to resolve that
question in order to land this commit -- even if we _do_
decide we need some kind of truncation, it will need to be done
in a different way, so we might as well just delette this code path for now.