coord: correctly stitch up tracing/otel context of peek stages
Before, we would hand the otel context of one stage to the next stage,
meaning the traces would be nested.

Now, we hand across a "root" otel context, meaning the different peek
stages are peers when it comes to traces. This makes them easier to look
at and makes folding of tracing views work nicer.