webhook: Move timestamp evaluation to when we receive the request (#23473)
This PR updates the logic for evaluating "now" when it's used in the
validation statement of a webhook. Instead of getting "now" when
validating a request, we get "now" when we receive the request. It's
possible for 10s of seconds to pass between receiving a request and when
it gets evaluated, if the Coordinator is blocked, this was occurring
during a run of the parallel-workload test.

It also updates the `CHECK` statement used in the parallel-workload test
to have a 10 second buffer, which is how we recommend users run webhooks
in production.

### Motivation

Fixes https://github.com/MaterializeInc/materialize/issues/23461

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered.
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a