Fix statement logging for subscribes in HTTP queries (incl. websocket). (#23534)
The protocol for statement logging of subscribes is as follows:

1. When the statement begins execution, an ID is generated (this is the
payload of the `ExecuteContextExtra`). This struct has a `Drop` impl
that logs an error (actually a soft_panic) if we forget to retire it.
2. This `ExecuteContextExtra` is handed back to the client (pgwire or
http) as part of the `ExecuteResponse` for subscribes.
3. Once the subscribe ends (either by finishing normally, erroring, or
being canceled) the client is supposed to send the `ExecuteContextExtra`
back to the coordinator for the statement log entry to be finalized.

This works fine in pgwire, but step 3 never got implemented in the HTTP
layer.

### Motivation

Fixes #21598 

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- Fixed a bug causing `SUBSCRIBE` statements issued via the web console
to never be marked as finished in `mz_activity_log`.