sql-parser: further improve error for trailing comma in SELECT list
Trailing commas in SELECT lists, e.g.:

    SELECT
       1,
       2,
    FROM ...

are common enough that it's worth a little bit of extra code to make the
error message as explicit as possible that it's a trailing comma
problem. This commit changes the error message from "expected
expression, but found reserved keyword: FROM" to "invalid trailing comma
in SELECT list".

Follow up to #23621/#23627.