*: make more soft-asserts logging
This commit enables logging for more soft-asserts. In general, we want
to default to make soft-asserts log by default and only keep them silent
when the overhead of performing the condition check would be too high in
production.

This commit takes a conservative approach and only switches to the
`_or_log` variants for soft-asserts that are not on the hot path and
have a cheap condition.