Add max persist_sink correction length metric
This commit adds a metric, mz_persist_sink_correction_max_len, that
captures the maximum number of updates seen in the correction buffer
per worker across all instances of persist_sink. The metric should
correlate with the memory spike produced by instances of persist_sink
during hydration, but not model it perfectly. This is because
concurrent (as opposed to sequential) hydration can create conditions
where the peaks introduced by multiple instances of persist_sink
reinforce each other. Under sequential hydration, this metric would
give us an upper bound on the length of the buffer that we would
expect to see.