storage: quiet a warning when using the nightly compiler
The nightly rust compiler seems to correctly be finding that this method
is unused and warning about it. I assume that it was used at some point
and then the caller went away. While we investigate whether the caller
is coming back or if we should delete it, except it from the lint.