Run persist peek in a timely worker
This is almost equivalent, but it has the nice side effect of waking up
the timely poll loop.

Previous versions of this code would see occasional high peek latencies,
even though the actual peek logic was fast. It turns out that this was
timely's step_or_park function parking... it had no idea that progress
had been made in our background task, so it would stay parked. Doing the
peek in a timely worker means that the peek will complete with a timely
step, waking the poll loop so we can pass on the results.