Move WASM crates to their own workspace.
Due to cache invalidation issues we've been excluding WASM crates from
the main Cargo workspace. This has an undesirable side-effect in that
dependabot will not track excluded crates.

This mitigates the issue by moving the WASM crates to their own
workspace, and configuring Dependabot to track that one, too. It also
updates our Python Cargo helper to be aware of wildcards in the exclude
field (as well as missing exclude fields).