persist-txn: enable in CI with "eager uppers"
The new persist-txn code has three settings:
- off: The previous implementation of tables. Each is a persist shard
  with an upper that is advanced at least once per second by group
  commit.
- eager: The persist-txn "WAL" (the txns shard) is used as the source of
  truth for what is committed to tables. This is denormalized into every
  table (data shard) whenever the txns shard upper advances, meaning
  every data shard's upper advances whenever the txns shard does. This
  means that all persist reads on tables can be served through the
  previous codepaths.
- lazy: The persist-txn "WAL" (the txns shard) is used as the source of
  truth for what is committed to tables. This is denormalized into the
  tables (data shards) _only when they change_. This means the CRDB
  traffic is no longer a function of the total number of tables, instead
  it's a function of how many tables are actually changed in a txn.
  However, this means that all reads of tables (data shards) need to go
  through persist-txn so that it can account for the lazy upper
  behavior.

This PR sets CI to "eager" as a necessary precondition for rolling
"eager" out to prod. We'd like to get all the way to "lazy" but eager is
enough to unblock the Pv2 Use Case Isolation work. Breaking the rollout
to prod into two steps also de-risks it because "off" -> "eager" only
changes write paths and then "eager" -> "lazy" only changes read paths.
We'll later switch CI to "lazy" before rolling that out to prod.

The selected behavior of this feature is set on envd startup and not
again. To select a new behavior, envd needs to be restarted (which in
turn intentionally restarts clusterd). In order of decreasing
priority/override:

- The value of the --enable-persist-txn-tables cli flag, if any. This is
  possibly vestigial given the new default_system_parameters sniffing in
  this commit, but I'd like to keep it around for now in case it helps
  us write tests.
- The "default_system_parameters" flag/env, which is how we hook it up
  to the usual CI override list in __init__.py.
- The most recent value, if any, set for the `enable_persist_txn_tables`
  system var before restart, which is how we hook it up to Launch
  Darkly for the staging/prod rollout.